Fix compiler warnings on 64-bit Windows
GCC reports various instances of

warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]

and MSVC equivalently

warning C4312: 'type cast': conversion from 'int' to 'void *' of greater size
warning C4311: 'type cast': pointer truncation from 'void *' to 'long'

in ECPG test files.  This is because void* and long are cast back and
forth, but on 64-bit Windows, these have different sizes.  Fix by
using intptr_t instead.

The code actually worked fine because the integer values in use are
all small.  So this is just to get the test code to compile warning-free.

This change is simplified by having made stdint.h required (commit
957338418b69e9774ccc1bab59f765a62f0aa6f9).  Before this it would have
been more complicated because the ecpg test source files don't use the
full pg_config.h.

Discussion: https://www.postgresql.org/message-id/flat/5d398bbb-262a-5fed-d839-d0e5cff3c0d7%402ndquadrant.com

+++ src/interfaces/ecpg/test/expected/thread-alloc.c
@@ -7,6 +7,7 @@
 #define ECPGdebug(X,Y) ECPGdebug((X)+100,(Y))
 
 #line 1 "alloc.pgc"
+#include <stdint.h>
 #include <stdlib.h>
 #include "ecpg_config.h"
 
@@ -100,7 +101,7 @@ struct sqlca_t *ECPGget_sqlca(void);
 
 #endif
 
-#line 25 "alloc.pgc"
+#line 26 "alloc.pgc"
 
 
 #line 1 "regression.h"
@@ -110,14 +111,14 @@ struct sqlca_t *ECPGget_sqlca(void);
 
 
 
-#line 26 "alloc.pgc"
+#line 27 "alloc.pgc"
 
 
 /* exec sql whenever sqlerror  sqlprint ; */
-#line 28 "alloc.pgc"
+#line 29 "alloc.pgc"
 
 /* exec sql whenever not found  sqlprint ; */
-#line 29 "alloc.pgc"
+#line 30 "alloc.pgc"
 
 
 #ifdef WIN32
@@ -133,54 +134,54 @@ static void* fn(void* arg)
 	 
 	   
 	
-#line 40 "alloc.pgc"
+#line 41 "alloc.pgc"
  int value ;
  
-#line 41 "alloc.pgc"
+#line 42 "alloc.pgc"
  char name [ 100 ] ;
  
-#line 42 "alloc.pgc"
+#line 43 "alloc.pgc"
  char ** r = NULL ;
 /* exec sql end declare section */
-#line 43 "alloc.pgc"
+#line 44 "alloc.pgc"
 
 
-	value = (long)arg;
+	value = (intptr_t) arg;
 	sprintf(name, "Connection: %d", value);
 
 	{ ECPGconnect(__LINE__, 0, "ecpg1_regression" , NULL, NULL , name, 0); 
-#line 48 "alloc.pgc"
+#line 49 "alloc.pgc"
 
 if (sqlca.sqlcode < 0) sqlprint();}
-#line 48 "alloc.pgc"
+#line 49 "alloc.pgc"
 
 	{ ECPGsetcommit(__LINE__, "on", NULL);
-#line 49 "alloc.pgc"
+#line 50 "alloc.pgc"
 
 if (sqlca.sqlcode < 0) sqlprint();}
-#line 49 "alloc.pgc"
+#line 50 "alloc.pgc"
 
 	for (i = 1; i <= REPEATS; ++i)
 	{
 		{ ECPGdo(__LINE__, 0, 1, NULL, 0, ECPGst_normal, "select relname from pg_class where relname = 'pg_class'", ECPGt_EOIT, 
 	ECPGt_char,&(r),(long)0,(long)0,(1)*sizeof(char), 
 	ECPGt_NO_INDICATOR, NULL , 0L, 0L, 0L, ECPGt_EORT);
-#line 52 "alloc.pgc"
+#line 53 "alloc.pgc"
 
 if (sqlca.sqlcode == ECPG_NOT_FOUND) sqlprint();
-#line 52 "alloc.pgc"
+#line 53 "alloc.pgc"
 
 if (sqlca.sqlcode < 0) sqlprint();}
-#line 52 "alloc.pgc"
+#line 53 "alloc.pgc"
 
 		free(r);
 		r = NULL;
 	}
 	{ ECPGdisconnect(__LINE__, name);
-#line 56 "alloc.pgc"
+#line 57 "alloc.pgc"
 
 if (sqlca.sqlcode < 0) sqlprint();}
-#line 56 "alloc.pgc"
+#line 57 "alloc.pgc"
 
 
 	return 0;
@@ -188,7 +189,7 @@ if (sqlca.sqlcode < 0) sqlprint();}
 
 int main ()
 {
-	int i;
+	intptr_t i;
 #ifdef WIN32
 	HANDLE threads[THREADS];
 #else
@@ -207,7 +208,7 @@ int main ()
 		CloseHandle(threads[i]);
 #else
 	for (i = 0; i < THREADS; ++i)
-		pthread_create(&threads[i], NULL, fn, (void *) (long) i);
+		pthread_create(&threads[i], NULL, fn, (void *) i);
 	for (i = 0; i < THREADS; ++i)
 		pthread_join(threads[i], NULL);
 #endif
